% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_rss.R
\name{simple_rss.default}
\alias{simple_rss.default}
\title{\code{simple_rss} simple_rss default, which checks for availability
of the "residuals" generic for the class}
\usage{
\method{simple_rss}{default}(x, ...)
}
\arguments{
\item{x}{an object, for which to calculate the RSS}

\item{...}{other arguments passed to specific methods}
}
\value{
a numeric
}
\description{
\code{simple_rss} simple_rss default, which checks for availability
of the "residuals" generic for the class
}
