% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_rss.R
\name{simple_rss.randomForest}
\alias{simple_rss.randomForest}
\title{\code{simple_rss} simple_rss for Random Forests.
This method calculates the residuals manually, then calculates the RSS.}
\usage{
\method{simple_rss}{randomForest}(x, ...)
}
\arguments{
\item{x}{an object, for which to calculate the RSS}

\item{...}{other arguments passed to specific methods}
}
\value{
a numeric
}
\description{
\code{simple_rss} simple_rss for Random Forests.
This method calculates the residuals manually, then calculates the RSS.
}
