% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_rss.R
\name{simple_rss}
\alias{simple_rss}
\title{Calculate RSS}
\usage{
simple_rss(x, ...)
}
\arguments{
\item{x}{an object, for which to calculate the RSS}

\item{...}{other arguments passed to specific methods}
}
\value{
a numeric
}
\description{
Calculate the Residual Sum of Squares (RSS) for different model types
for many statistical models, one is interested in the accuracy
of fit of the data, often quantified by the RSS (residual sum of squares)
For some of these methods, the residuals function has already been defined
(such as for Recursive Partitioning Trees). However, for many it has not been defined
here we implement a default simple_rss method to calculate the RSS when the residuals
function is available, but use S3 to make it easily extensive when residuals is not available.
}
\examples{
library(randomForest)
data(iris)
rf <- randomForest(Sepal.Length ~ ., data=iris, importance=TRUE,
                   proximity=TRUE)
simple_rss(rf)
# 20.65949
}
\seealso{
\url{https://en.wikipedia.org/wiki/Residual_sum_of_squares}
}
